#75 - Split up Virus component
