+++ apps/lundin/src/app/modules/virus/virus-game/virus-game.component.html
@@ -1,0 +1,9 @@
+<div class="game-container">
+	<div *ngFor="let position of positions"
+		class="game-field"
+		(click)="select(position.x, position.y)"
+		[style.backgroundColor]="colorFor(position)"
+		[class.selected]="isSelected(position)">
+	</div>
+</div>
+<div class="message">{{message}}</div>
