+++ apps/lundin/src/app/modules/virus/virus.component.scss
@@ -1,21 +1,3 @@
 :host {
 	height: 100%;
 }
-
-.game-container {
-	display: grid;
-	grid-template-columns: repeat(8, 2rem);
-	grid-template-rows: repeat(8, 2rem);
-}
-
-.game-field {
-	line-height: 2rem;
-	text-align: center;
-	border: 1px solid black;
-}
-
-.selected{
-	outline: 3px solid white;
-	outline-style: solid;
-	outline-offset: -4px;
-}
